stripe payment , payment deatils store update

+++ database/migrations/2020_08_14_060813_create_user_payments_table.php
@@ -19,7 +19,7 @@ class CreateUserPaymentsTable extends Migration
             $table->string('street');
             $table->string('city');
             $table->string('country');
-            $table->integer('card_number');
+            $table->bigInteger('card_number');
             $table->integer('exp_month');
             $table->integer('exp_year');
             $table->integer('cvv');
